cambios generales

+++ Modules/Tarea/Database/Migrations/2022_08_08_164019_create_tareas_table.php
@@ -17,7 +17,7 @@ class CreateTareasTable extends Migration
             $table->id();
             $table->string('titulo', 100)->nullable();
             $table->string('descripcion');
-            $table->integer('personal_id');
+            $table->unsignedBigInteger('personal_id');
             $table->integer('porcentaje');
             $table->integer('estado_tarea');
             $table->string('fecha_inicio');
